# openapi.yml
openapi: 3.0.3
info:
  title: Space Marine Management API
  version: 1.0.0
  description: API for managing Space Marines, their Coordinates, and Chapters with full CRUD, pagination support, and special operations.
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
...
  /space-marines:
    get:
      summary: Get all Space Marines (paginated)
      operationId: getAllSpaceMarines
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
      responses:
        "200":
          description: Paginated list of Space Marines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSpaceMarineResponse"
    post:
      summary: Create a new Space Marine
      operationId: createSpaceMarine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceMarineCreateRequest"
      responses:
        "201":
          description: Space Marine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceMarine"
        "400":
          description: Invalid input
  /space-marines/{id}:
    get:
      summary: Get a Space Marine by ID
      operationId: getSpaceMarineById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        "200":
          description: Space Marine found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceMarine"
        "404":
          description: Space Marine not found

    put:
      summary: Update a Space Marine by ID
      operationId: updateSpaceMarine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceMarineUpdateRequest"
      responses:
        "200":
          description: Space Marine updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceMarine"
        "400":
          description: Invalid input
        "404":
          description: Space Marine not found

    delete:
      summary: Delete a Space Marine by ID
      operationId: deleteSpaceMarine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        "204":
          description: Space Marine deleted
        "404":
          description: Space Marine not found
...
components:
  schemas:
...
    PaginatedSpaceMarineResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SpaceMarine"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
...
    SpaceMarine:
      type: object
      required:
        - id
        - name
        - coordinatesId
        - creationDate
        - chapterId
        - health
        - weaponType
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          description: Unique auto-generated ID (> 0)
        name:
          type: string
          minLength: 1
          description: Non-empty, non-null name
        coordinatesId:
          type: integer
          format: int64
          minimum: 1
          description: ID of associated Coordinates
        creationDate:
          type: string
          format: date-time
          description: Auto-generated creation timestamp
        chapterId:
          type: integer
          format: int64
          minimum: 1
          description: ID of associated Chapter
        health:
          type: integer
          format: int64
          minimum: 1
          description: Health must be > 0
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
          description: Astartes category
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]
          description: Type of weapon used by the Space Marine
    SpaceMarineCreateRequest:
      type: object
      required:
        - name
        - coordinatesId
        - chapterId
        - health
        - weaponType
      properties:
        name:
          type: string
          minLength: 1
        coordinatesId:
          type: integer
          format: int64
          minimum: 1
        chapterId:
          type: integer
          format: int64
          minimum: 1
        health:
          type: integer
          format: int64
          minimum: 1
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]
    SpaceMarineUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        coordinatesId:
          type: integer
          format: int64
          minimum: 1
        chapterId:
          type: integer
          format: int64
          minimum: 1
        health:
          type: integer
          format: int64
          minimum: 1
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]

    # === COORDINATES ===
...